+++ src/entities/Massengutartikel.java
@@ -1,9 +1,11 @@
 package entities;
 
+import java.math.BigDecimal;
+
 public class Massengutartikel extends Artikel {
     private int packungGroesse;
 
-    public Massengutartikel(int artikelID, String bezeichnung, float preis, int packungGroesse) {
+    public Massengutartikel(int artikelID, String bezeichnung, BigDecimal preis, int packungGroesse) {
         super(artikelID, bezeichnung, preis);
         this.packungGroesse = packungGroesse;
     }
